:bug: add word alternatives to speech to text
