+++ examples/java/com/ibm/watson/developer_cloud/speech_to_text/v1/MicrophoneWithWebSocketsExample.java
@@ -1,9 +1,26 @@
+/*
+ * Copyright 2015 IBM Corp. All Rights Reserved.
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License"); you may not use this file except
+ * in compliance with the License. You may obtain a copy of the License at
+ *
+ * http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software distributed under the License
+ * is distributed on an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express
+ * or implied. See the License for the specific language governing permissions and limitations under
+ * the License.
+ */
 package com.ibm.watson.developer_cloud.speech_to_text.v1;
 
-import java.io.FileInputStream;
 import java.io.FileNotFoundException;
-import java.util.concurrent.CountDownLatch;
-import java.util.concurrent.TimeUnit;
+
+import javax.sound.sampled.AudioFormat;
+import javax.sound.sampled.AudioInputStream;
+import javax.sound.sampled.AudioSystem;
+import javax.sound.sampled.DataLine;
+import javax.sound.sampled.LineUnavailableException;
+import javax.sound.sampled.TargetDataLine;
 
 import com.ibm.watson.developer_cloud.http.HttpMediaType;
 import com.ibm.watson.developer_cloud.speech_to_text.v1.model.RecognizeOptions;
@@ -14,7 +31,14 @@ import com.ibm.watson.developer_cloud.speech_to_text.v1.websocket.BaseRecognizeC
  * Recognize microphone input speech continuously using WebSockets.
  */
 public class MicrophoneWithWebSocketsExample {
-  public static void main(String[] args) throws FileNotFoundException, InterruptedException {
+  
+  /**
+   * The main method.
+   *
+   * @param args the arguments
+   * @throws Exception the exception
+   */
+  public static void main(final String[] args) throws Exception {
     SpeechToText service = new SpeechToText();
     service.setUsernameAndPassword("<username>", "<password>");
